Correctly check permissions before mutating a comment
Previously the api depended on the client to provide id of the author of
a comment to be check against current user's id to allow for mutating a
comment. That is incorrect as a user who isn't the actual author a comment
can simply put their id as the author of the comment and it would be valid
when it is checked against their id in the jwt and then they would be able
to mutate the comment. This fix removes the dependence on the client and
uses the DB to check for proper user permissions.