Correctly check permissions before mutating a post
Previously the api depended on the client to provide id of the author of
a post to be check against current user's id to allow for mutating a
post. That is incorrect as a user who isn't the actual author a post
can simply put their id as the author of the post and it would be valid
when it is checked against their id in the jwt and then they would be able
to mutate the post. This fix removes the dependence on the client and
uses the DB to check for proper user permissions.